backend/requirements.txt
fastapi
uvicorn
python-dotenv
feedparser
requests
tqdm
pydantic
reportlab
sentence-transformers
faiss-cpu
openai
